Fix types of arguments to `GetSubtreeRoots` and `putSaplingSubtreeRoots`

+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/TypesafeBackend.kt
@@ -82,7 +82,7 @@ internal interface TypesafeBackend {
      */
     @Throws(RuntimeException::class)
     suspend fun putSaplingSubtreeRoots(
-        startIndex: Long,
+        startIndex: UInt,
         roots: List<SubtreeRoot>,
     )
